refactor: add user controller for work with user requests
