+++ package.json
@@ -35,6 +35,6 @@
   },
   "dependencies": {
     "dotenv": "^16.4.5",
-    "uuid": "^10.0.0"
+    "uuidv4": "^6.2.13"
   }
 }
